refactor: 롤링페이퍼 결과 DTO 위치 조정
