+++ src/main/kotlin/com/team2/server/rollingpaper/controller/RollingPaperToppingApi.kt
@@ -3,7 +3,7 @@ package com.team2.server.rollingpaper.controller
 import com.team2.server.common.web.ApiResponse
 import com.team2.server.common.web.swagger.InternalServerErrorResponse
 import com.team2.server.common.web.swagger.OptionalAuth
-import com.team2.server.rollingpaper.application.dto.RollingPaperToppingResult
+import com.team2.server.rollingpaper.dto.RollingPaperToppingResult
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.security.SecurityRequirement
 import io.swagger.v3.oas.annotations.tags.Tag
