+++ src/main/kotlin/com/team2/server/rollingpaper/controller/RollingPaperToppingController.kt
@@ -1,7 +1,7 @@
 package com.team2.server.rollingpaper.controller
 
 import com.team2.server.common.web.ApiResponse
-import com.team2.server.rollingpaper.application.dto.RollingPaperToppingResult
+import com.team2.server.rollingpaper.dto.RollingPaperToppingResult
 import com.team2.server.rollingpaper.usecase.GetRollingPaperToppingsUseCase
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
