+++ src/main/kotlin/com/team2/server/rollingpaper/dto/RollingPaperToppingResult.kt
@@ -1,4 +1,4 @@
-package com.team2.server.rollingpaper.application.dto
+package com.team2.server.rollingpaper.dto
 
 import io.swagger.v3.oas.annotations.media.Schema
